1) Added support to compute super jacobian from base jacobian when n_base_iter == -1 2) Fixed bug in chain rule for Log transformation 3) Added limit to number of times the super iteration will try to freeze parameters when computing the jacobian_forward

+++ pest++/pestpp_common/eigen_tools.cpp
@@ -74,7 +74,7 @@ Eigen::VectorXd transformable_2_egien_vec(const Transformable &data, vector<stri
 }
 
 
-void matrix_del_cols(Eigen::SparseMatrix<double> &mat, const vector<int> &col_id_vec)
+void matrix_del_cols(Eigen::SparseMatrix<double> &mat, const vector<size_t> &col_id_vec)
 {
 	if (!col_id_vec.empty())
 	{
